Move WorkloadWatcher to traffic/cmd/manager
This commit just moves the code. There's no change in logic.
