+++ cmd/traffic/cmd/manager/state/intercept_test.go
@@ -16,7 +16,6 @@ import (
 	"github.com/telepresenceio/telepresence/v2/pkg/agentconfig"
 	"github.com/telepresenceio/telepresence/v2/pkg/log"
 	"github.com/telepresenceio/telepresence/v2/pkg/tunnel"
-	"github.com/telepresenceio/telepresence/v2/pkg/workload"
 )
 
 // TestAllowGlobalIntercepts_ValidationLogic tests the validation logic
@@ -110,7 +109,7 @@ func TestAllowGlobalIntercepts_ValidationLogic(t *testing.T) {
 				intercepts:       watchable.NewMap[string, *Intercept](interceptEqual, time.Millisecond),
 				agents:           watchable.NewMap[tunnel.SessionID, *AgentSession](agentsEqual, time.Millisecond),
 				clients:          xsync.NewMap[tunnel.SessionID, *ClientSession](),
-				workloadWatchers: xsync.NewMap[string, workload.Watcher](),
+				workloadWatchers: xsync.NewMap[string, Watcher](),
 				timedLogLevel:    log.NewTimedLevel("debug", log.SetLevel),
 				llSubs:           newLoglevelSubscribers(),
 			}
@@ -169,7 +168,7 @@ func TestAllowGlobalIntercepts_ErrorMessage(t *testing.T) {
 		intercepts:       watchable.NewMap[string, *Intercept](interceptEqual, time.Millisecond),
 		agents:           watchable.NewMap[tunnel.SessionID, *AgentSession](agentsEqual, time.Millisecond),
 		clients:          xsync.NewMap[tunnel.SessionID, *ClientSession](),
-		workloadWatchers: xsync.NewMap[string, workload.Watcher](),
+		workloadWatchers: xsync.NewMap[string, Watcher](),
 		timedLogLevel:    log.NewTimedLevel("debug", log.SetLevel),
 		llSubs:           newLoglevelSubscribers(),
 	}
@@ -259,7 +258,7 @@ func TestAllowGlobalIntercepts_DefaultBehavior(t *testing.T) {
 		intercepts:       watchable.NewMap[string, *Intercept](interceptEqual, time.Millisecond),
 		agents:           watchable.NewMap[tunnel.SessionID, *AgentSession](agentsEqual, time.Millisecond),
 		clients:          xsync.NewMap[tunnel.SessionID, *ClientSession](),
-		workloadWatchers: xsync.NewMap[string, workload.Watcher](),
+		workloadWatchers: xsync.NewMap[string, Watcher](),
 		timedLogLevel:    log.NewTimedLevel("debug", log.SetLevel),
 		llSubs:           newLoglevelSubscribers(),
 	}
